pyproject.toml
[project]
name = "ipt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "ipykernel>=7.2.0",
    "loguru>=0.7.3",
    "pandas>=3.0.2",
    "playwright>=1.58.0",
    "protobuf>=7.34.1",
    "pyabsa>=2.4.2",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "sentencepiece>=0.2.1",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformers>=4.44.0,<5.0",
]
